Files are renamed to comply with the Kebab Case convention
