+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import '@/styles/globals.css';
 import type { AppProps } from 'next/app';
-import { MainLayout } from '../components/layouts/MainLayout';
+import { MainLayout } from '../components/layouts/main-layout';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
